move clientsHydrated to clients store
