+++ webapp/store/actions/appActions.js
@@ -21,13 +21,6 @@ export function getWindowInfo() {
   };
 }
 
-export function setClientsHydrated(hydrated = false) {
-  return {
-    type: types.CLIENTS_HYDRATED,
-    payload: hydrated
-  };
-}
-
 export default {
   getWindowInfo
 };
