+++ webapp/store/constants/app.js
@@ -2,4 +2,3 @@ export const SET_WINDOW = 'SET_WINDOW';
 export const STATE_REFRESHED = 'STATE_REFRESHED';
 export const RESET_STATE = 'RESET_STATE';
 export const APP_LOADED = 'APP_LOADED';
-export const CLIENTS_HYDRATED = 'CLIENTS_HYDRATED';
