+++ webapp/store/constants/clients.js
@@ -5,3 +5,4 @@ export const UPDATE_CLIENT = 'UPDATE_CLIENT';
 export const ADD_CLIENT = 'ADD_CLIENT';
 export const REMOVE_CLIENT = 'REMOVE_CLIENT';
 export const CLEAR_CURRENT_CLIENT = 'CLEAR_CURRENT_CLIENT';
+export const CLIENTS_HYDRATED = 'CLIENTS_HYDRATED';
